fix(data): update TopDiscountedDukanDto discount type from Int to Double

+++ dukan_data/src/commonMain/kotlin/net/thechance/mena/dukan/data/dto/dukan/TopDiscountedDukanDto.kt
@@ -13,7 +13,7 @@ data class TopDiscountedDukanDto(
     val imageUrl: String,
 
     @SerialName("discount")
-    val discount: Int,
+    val discount: Double,
 
     @SerialName("id")
     val id: Uuid
